requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-texttospeech==2.16.5
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-cloud-tasks==2.16.3
firebase-admin==6.5.0
trafilatura==1.9.0
beautifulsoup4==4.12.3
feedgen==1.0.0
pydub==0.25.1
requests==2.32.3
